feat: If the extension is (re)loaded and there are no subscriptions, subscribe to default contributors
chore: Keep await rehydration
