+++ src/app/background/sagas/install.ts
@@ -1,5 +1,5 @@
 import { SagaIterator } from 'redux-saga';
-import { takeLatest, select, put, call, all } from 'redux-saga/effects';
+import { takeLatest, select, put, call } from 'redux-saga/effects';
 import { captureException } from 'app/utils/sentry';
 import openOptions from 'webext/openOptionsTab';
 import {
@@ -11,18 +11,11 @@ import { isOnboardingRequired } from 'app/background/selectors';
 import { getInstallationDate } from 'app/background/selectors/installationDetails';
 import { InstallationDetails } from 'app/lmem/installation';
 import { version } from '../../../../package.json';
-import { subscribe } from 'app/actions';
-import { preselectedContributorIds } from 'app/lmemContributors';
-import { loginSaga } from './user.saga';
-import awaitRehydrationSaga from './lib/awaitRehydration.saga';
-
 export function* installedSaga({
   payload: { installedDetails }
 }: InstalledAction): SagaIterator {
   try {
     const datetime = yield select(getInstallationDate);
-
-    yield call(loginSaga);
 
     // @todo why not use this function to get the current version ?
     // const version = chrome.runtime.getManifest().version;
@@ -34,18 +27,10 @@ export function* installedSaga({
       updatedAt: new Date()
     };
 
-    yield call(awaitRehydrationSaga);
-
     yield put(updateInstallationDetails(installationDetails, false));
 
     const { reason } = installedDetails;
     if (reason === 'install') {
-      yield all(
-        preselectedContributorIds.map(contributorId =>
-          put(subscribe(contributorId))
-        )
-      );
-
       const onboardingRequired = yield select(isOnboardingRequired);
       if (onboardingRequired) {
         yield call(openOptions, '/onboarding');
